Split out more requests + responses
