+++ osu.Game/Online/API/Requests/GetRoomScoresRequest.cs
@@ -1,0 +1,20 @@
+// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+using System.Collections.Generic;
+using osu.Game.Online.API.Requests.Responses;
+
+namespace osu.Game.Online.API.Requests
+{
+    public class GetRoomScoresRequest : APIRequest<List<APIRoomScoreInfo>>
+    {
+        private readonly int roomId;
+
+        public GetRoomScoresRequest(int roomId)
+        {
+            this.roomId = roomId;
+        }
+
+        protected override string Target => $@"rooms/{roomId}/leaderboard";
+    }
+}
